+++ views/js/cameraViewRunner.js
@@ -159,6 +159,9 @@ function grabElements(printer) {
             iconBedA: document.getElementById("bedA-" + printer._id),
             iconTool0A: document.getElementById("tool0A-" + printer._id),
             iconTool0T: document.getElementById("tool0T-" + printer._id),
+            extraInfo: document.getElementById("extraInfo-" + printer._id),
+            timeRemaining: document.getElementById("timeRemaining-" + printer._id),
+            eta: document.getElementById("eta-" + printer._id),
         };
         elems[printer._id] = printerElemens;
         return elems[printer._id];
@@ -168,7 +171,29 @@ function grabElements(printer) {
 function updateState(printers, clientSettings) {
     printers.forEach((printer) => {
         let elements = grabElements(printer);
+        if(clientSettings.extraInfo){
+            if(elements.extraInfo.classList.contains("d-none")){
+                elements.extraInfo.classList.remove("d-none");
+            }
 
+            if(typeof printer.progress != 'undefined' && printer.progress.printTimeLeft != 0){
+                let currentDate = new Date();
+                currentDate = currentDate.getTime();
+                let futureDateString = new Date(currentDate + printer.progress.printTimeLeft * 1000).toDateString()
+                let futureTimeString = new Date(currentDate + printer.progress.printTimeLeft * 1000).toTimeString()
+                futureTimeString = futureTimeString.substring(0, 8);
+                let dateComplete = futureDateString + ": " + futureTimeString;
+                elements.timeRemaining.innerHTML = `
+        ${Calc.generateTime(printer.progress.printTimeLeft)}
+      `;
+                elements.eta.innerHTML = dateComplete
+            }else{
+                elements.timeRemaining.innerHTML = `
+        ${Calc.generateTime(null)}
+      `;
+                elements.eta.innerHTML = "N/A"
+            }
+        }
         elements.camBackground.className = `card-body cameraContain ${printer.stateColour.category}`;
         if (typeof printer.job != "undefined" && printer.job.file.name != null) {
             elements.currentFile.setAttribute("title", printer.job.file.path)