+++ views/js/listViewRunner.js
@@ -140,7 +140,10 @@ function grabElements(printer) {
       iconBedT: document.getElementById("bedT-" + printer._id),
       iconBedA: document.getElementById("bedA-" + printer._id),
       iconTool0A: document.getElementById("tool0A-" + printer._id),
-      iconTool0T: document.getElementById("tool0T-" + printer._id)
+      iconTool0T: document.getElementById("tool0T-" + printer._id),
+      extraInfoCol: document.getElementById("extraInfoCol-" + printer._id),
+      extraInfoTitle: document.getElementById("extraInfoTitle"),
+      eta: document.getElementById("eta-" + printer._id),
     };
     elems[printer._id] = printerElemens;
     return elems[printer._id];
@@ -153,7 +156,26 @@ function updateState(printers, clientSettings, filamentProfiles, filamentManager
   printers.forEach(async printer => {
     let elements = grabElements(printer);
     //Set the data
+    if(clientSettings.extraInfo){
+      if(elements.extraInfoCol.classList.contains("d-none")){
+        elements.extraInfoCol.classList.remove("d-none");
+        elements.extraInfoTitle.classList.remove("d-none");
+      }
 
+      if(typeof printer.progress != 'undefined'){
+        let currentDate = new Date();
+        currentDate = currentDate.getTime();
+        let futureDateString = new Date(currentDate + printer.progress.printTimeLeft * 1000).toDateString()
+        let futureTimeString = new Date(currentDate + printer.progress.printTimeLeft * 1000).toTimeString()
+        futureTimeString = futureTimeString.substring(0, 8);
+        let dateComplete = futureDateString + ": " + futureTimeString;
+        elements.eta.innerHTML = dateComplete
+      }else{
+        elements.eta.innerHTML = "N/A"
+      }
+
+      elements.eta;
+    }
 
     if (typeof printer.job != "undefined" && printer.job.file.name != null) {
       elements.currentFile.setAttribute('title', printer.job.file.path)